admin/requirements.txt
flask==2.0.1
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0 